requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.11
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
